+++ rollup.config.js
@@ -19,7 +19,12 @@ const external = [
   'emojibase',
   'escape-html',
   'interweave',
+  'interweave-emoji',
   'lodash',
+  'lodash/camelCase',
+  'lodash/chunk',
+  'lodash/debounce',
+  'lodash/upperFirst',
   'prop-types',
   'react-window',
   'react',
